Backgrounds as union

+++ include/types_draw.h
@@ -2,20 +2,20 @@
 
 // Types common for several steps in the drawing process
 
-#include "types.h" // i32
+#include "types.h" // i16
 
 typedef struct {
-  i32 top;
-  i32 right;
-  i32 bottom;
-  i32 left;
+  i16 top;
+  i16 right;
+  i16 bottom;
+  i16 left;
 } Padding;
 
 typedef struct {
-  i32 top;
-  i32 right;
-  i32 bottom;
-  i32 left;
+  i16 top;
+  i16 right;
+  i16 bottom;
+  i16 left;
 } Border;
 
 #define C9_TYPES_DRAW
